+++ tests/test_config.py
@@ -4,8 +4,9 @@ import warnings
 
 import pytest
 from pydantic import ValidationError
+from pytest import MonkeyPatch
 
-from stac_fastapi.pgstac.config import PostgresSettings
+from stac_fastapi.pgstac.config import PostgresSettings, Settings
 
 
 async def test_pg_settings_with_env(monkeypatch):
@@ -74,3 +75,60 @@ async def test_pg_settings_attributes(monkeypatch):
                 postgres_dbname="pgstac",
                 _env_file=None,
             )
+
+
+@pytest.mark.parametrize(
+    "cors_origins",
+    [
+        "http://stac-fastapi-pgstac.test,http://stac-fastapi.test",
+        '["http://stac-fastapi-pgstac.test","http://stac-fastapi.test"]',
+    ],
+)
+def test_cors_origins(monkeypatch: MonkeyPatch, cors_origins: str) -> None:
+    monkeypatch.setenv(
+        "CORS_ORIGINS",
+        cors_origins,
+    )
+    settings = Settings()
+    assert settings.cors_origins == [
+        "http://stac-fastapi-pgstac.test",
+        "http://stac-fastapi.test",
+    ]
+
+
+@pytest.mark.parametrize(
+    "cors_methods",
+    [
+        "GET,POST",
+        '["GET","POST"]',
+    ],
+)
+def test_cors_methods(monkeypatch: MonkeyPatch, cors_methods: str) -> None:
+    monkeypatch.setenv(
+        "CORS_METHODS",
+        cors_methods,
+    )
+    settings = Settings()
+    assert settings.cors_methods == [
+        "GET",
+        "POST",
+    ]
+
+
+@pytest.mark.parametrize(
+    "cors_headers",
+    [
+        "Content-Type,X-Foo",
+        '["Content-Type","X-Foo"]',
+    ],
+)
+def test_cors_headers(monkeypatch: MonkeyPatch, cors_headers: str) -> None:
+    monkeypatch.setenv(
+        "CORS_HEADERS",
+        cors_headers,
+    )
+    settings = Settings()
+    assert settings.cors_headers == [
+        "Content-Type",
+        "X-Foo",
+    ]